+++ Auragraph/AGRenderObject.mm
@@ -298,7 +298,7 @@ void AGRenderObject::debug_renderBounds()
 #endif // DEBUG_BOUNDS
 }
 
-void AGRenderObject::drawGeometry(GLvertex3f geo[], unsigned long size, int kind)
+void AGRenderObject::drawGeometry(GLvertex3f geo[], int size, int kind)
 {
     AGGenericShader &shader = AGGenericShader::instance();
     
@@ -310,10 +310,10 @@ void AGRenderObject::drawGeometry(GLvertex3f geo[], unsigned long size, int kind
     glVertexAttribPointer(AGVertexAttribPosition, 3, GL_FLOAT, false, 0, geo);
     glEnableVertexAttribArray(AGVertexAttribPosition);
     
-    glDrawArrays(kind, 0, (int) size);
+    glDrawArrays(kind, 0, size);
 }
 
-void AGRenderObject::drawTriangleFan(GLvertex2f geo[], unsigned long size)
+void AGRenderObject::drawTriangleFan(GLvertex2f geo[], int size)
 {
     AGGenericShader &shader = AGGenericShader::instance();
     
@@ -325,10 +325,10 @@ void AGRenderObject::drawTriangleFan(GLvertex2f geo[], unsigned long size)
     glVertexAttribPointer(AGVertexAttribPosition, 2, GL_FLOAT, false, 0, geo);
     glEnableVertexAttribArray(AGVertexAttribPosition);
     
-    glDrawArrays(GL_TRIANGLE_FAN, 0, (int) size);
+    glDrawArrays(GL_TRIANGLE_FAN, 0, size);
 }
 
-void AGRenderObject::drawTriangleFan(GLvertex3f geo[], unsigned long size)
+void AGRenderObject::drawTriangleFan(GLvertex3f geo[], int size)
 {
     AGGenericShader &shader = AGGenericShader::instance();
     
@@ -340,10 +340,10 @@ void AGRenderObject::drawTriangleFan(GLvertex3f geo[], unsigned long size)
     glVertexAttribPointer(AGVertexAttribPosition, 3, GL_FLOAT, false, 0, geo);
     glEnableVertexAttribArray(AGVertexAttribPosition);
     
-    glDrawArrays(GL_TRIANGLE_FAN, 0, (int) size);
+    glDrawArrays(GL_TRIANGLE_FAN, 0, size);
 }
 
-void AGRenderObject::drawTriangleFan(GLvertex3f geo[], unsigned long size, const GLKMatrix4 &xform)
+void AGRenderObject::drawTriangleFan(GLvertex3f geo[], int size, const GLKMatrix4 &xform)
 {
     AGGenericShader &shader = AGGenericShader::instance();
     
@@ -356,10 +356,10 @@ void AGRenderObject::drawTriangleFan(GLvertex3f geo[], unsigned long size, const
     glVertexAttribPointer(AGVertexAttribPosition, 3, GL_FLOAT, false, 0, geo);
     glEnableVertexAttribArray(AGVertexAttribPosition);
     
-    glDrawArrays(GL_TRIANGLE_FAN, 0, (int) size);
+    glDrawArrays(GL_TRIANGLE_FAN, 0, size);
 }
 
-void AGRenderObject::drawTriangleFan(AGGenericShader &shader, GLvertex3f geo[], unsigned long size, const GLKMatrix4 &xform)
+void AGRenderObject::drawTriangleFan(AGGenericShader &shader, GLvertex3f geo[], int size, const GLKMatrix4 &xform)
 {
 //    shader.useProgram();
     
@@ -370,10 +370,10 @@ void AGRenderObject::drawTriangleFan(AGGenericShader &shader, GLvertex3f geo[], 
     glVertexAttribPointer(AGVertexAttribPosition, 3, GL_FLOAT, false, 0, geo);
     glEnableVertexAttribArray(AGVertexAttribPosition);
     
-    glDrawArrays(GL_TRIANGLE_FAN, 0, (int) size);
+    glDrawArrays(GL_TRIANGLE_FAN, 0, size);
 }
 
-void AGRenderObject::drawLineLoop(GLvertex2f geo[], unsigned long size)
+void AGRenderObject::drawLineLoop(GLvertex2f geo[], int size)
 {
     AGGenericShader &shader = AGGenericShader::instance();
     
@@ -385,10 +385,10 @@ void AGRenderObject::drawLineLoop(GLvertex2f geo[], unsigned long size)
     glVertexAttribPointer(AGVertexAttribPosition, 2, GL_FLOAT, false, 0, geo);
     glEnableVertexAttribArray(AGVertexAttribPosition);
     
-    glDrawArrays(GL_LINE_LOOP, 0, (int) size);
+    glDrawArrays(GL_LINE_LOOP, 0, size);
 }
 
-void AGRenderObject::drawLineLoop(GLvertex3f geo[], unsigned long size)
+void AGRenderObject::drawLineLoop(GLvertex3f geo[], int size)
 {
     AGGenericShader &shader = AGGenericShader::instance();
     
@@ -400,10 +400,10 @@ void AGRenderObject::drawLineLoop(GLvertex3f geo[], unsigned long size)
     glVertexAttribPointer(AGVertexAttribPosition, 3, GL_FLOAT, false, 0, geo);
     glEnableVertexAttribArray(AGVertexAttribPosition);
     
-    glDrawArrays(GL_LINE_LOOP, 0, (int) size);
+    glDrawArrays(GL_LINE_LOOP, 0, size);
 }
 
-void AGRenderObject::drawLineLoop(GLvertex3f geo[], unsigned long size, const GLKMatrix4 &xform)
+void AGRenderObject::drawLineLoop(GLvertex3f geo[], int size, const GLKMatrix4 &xform)
 {
     AGGenericShader &shader = AGGenericShader::instance();
     
@@ -416,10 +416,10 @@ void AGRenderObject::drawLineLoop(GLvertex3f geo[], unsigned long size, const GL
     glVertexAttribPointer(AGVertexAttribPosition, 3, GL_FLOAT, false, 0, geo);
     glEnableVertexAttribArray(AGVertexAttribPosition);
     
-    glDrawArrays(GL_LINE_LOOP, 0, (int) size);
+    glDrawArrays(GL_LINE_LOOP, 0, size);
 }
 
-void AGRenderObject::drawLineStrip(GLvertex2f geo[], unsigned long size)
+void AGRenderObject::drawLineStrip(GLvertex2f geo[], int size)
 {
     AGGenericShader &shader = AGGenericShader::instance();
     
@@ -431,10 +431,10 @@ void AGRenderObject::drawLineStrip(GLvertex2f geo[], unsigned long size)
     glVertexAttribPointer(AGVertexAttribPosition, 2, GL_FLOAT, false, 0, geo);
     glEnableVertexAttribArray(AGVertexAttribPosition);
     
-    glDrawArrays(GL_LINE_STRIP, 0, (int) size);
+    glDrawArrays(GL_LINE_STRIP, 0, size);
 }
 
-void AGRenderObject::drawLineStrip(GLvertex2f geo[], unsigned long size, const GLKMatrix4 &xform)
+void AGRenderObject::drawLineStrip(GLvertex2f geo[], int size, const GLKMatrix4 &xform)
 {
     AGGenericShader &shader = AGGenericShader::instance();
     
@@ -447,10 +447,10 @@ void AGRenderObject::drawLineStrip(GLvertex2f geo[], unsigned long size, const G
     glVertexAttribPointer(AGVertexAttribPosition, 2, GL_FLOAT, false, 0, geo);
     glEnableVertexAttribArray(AGVertexAttribPosition);
     
-    glDrawArrays(GL_LINE_STRIP, 0, (int) size);
+    glDrawArrays(GL_LINE_STRIP, 0, size);
 }
 
-void AGRenderObject::drawLineStrip(AGGenericShader &shader, GLvertex2f geo[], unsigned long size, const GLKMatrix4 &xform)
+void AGRenderObject::drawLineStrip(AGGenericShader &shader, GLvertex2f geo[], int size, const GLKMatrix4 &xform)
 {
 //    shader.useProgram();
     
@@ -461,10 +461,10 @@ void AGRenderObject::drawLineStrip(AGGenericShader &shader, GLvertex2f geo[], un
     glVertexAttribPointer(AGVertexAttribPosition, 2, GL_FLOAT, false, 0, geo);
     glEnableVertexAttribArray(AGVertexAttribPosition);
     
-    glDrawArrays(GL_LINE_STRIP, 0, (int) size);
+    glDrawArrays(GL_LINE_STRIP, 0, size);
 }
 
-void AGRenderObject::drawLineStrip(GLvertex3f geo[], unsigned long size)
+void AGRenderObject::drawLineStrip(GLvertex3f geo[], int size)
 {
     AGGenericShader &shader = AGGenericShader::instance();
     
@@ -476,10 +476,10 @@ void AGRenderObject::drawLineStrip(GLvertex3f geo[], unsigned long size)
     glVertexAttribPointer(AGVertexAttribPosition, 3, GL_FLOAT, false, 0, geo);
     glEnableVertexAttribArray(AGVertexAttribPosition);
     
-    glDrawArrays(GL_LINE_STRIP, 0, (int) size);
+    glDrawArrays(GL_LINE_STRIP, 0, size);
 }
 
-void AGRenderObject::drawWaveform(float waveform[], unsigned long size, GLvertex2f from, GLvertex2f to, float gain, float yScale)
+void AGRenderObject::drawWaveform(float waveform[], int size, GLvertex2f from, GLvertex2f to, float gain, float yScale)
 {
     GLvertex2f vec = (to - from);
     
@@ -524,7 +524,7 @@ void AGRenderObject::drawWaveform(float waveform[], unsigned long size, GLvertex
     
     glLineWidth(2.0);
     
-    glDrawArrays(GL_LINE_STRIP, 0, (int) size);
+    glDrawArrays(GL_LINE_STRIP, 0, size);
     
     glEnableVertexAttribArray(AGVertexAttribPosition);
 }
